Setup timeout monitor for all CI builds
